Move main module to pywhip file
